Factor out deferred resolvable creation
